backend/requirements.txt
cat > requirements.txt << 'EOF'
fastapi==0.104.1
uvicorn==0.24.0
ultralytics==8.4.37
opencv-python-headless==4.8.1.78
python-multipart==0.0.6
pillow==10.1.0
numpy==1.24.3
EOF
